requirements.txt
streamlit>=1.44.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
streamlit-webrtc>=0.47.1
requests>=2.31.0
openai>=1.0.0
googletrans>=3.1.0a0
PyMySQL>=1.1.0 